--- a/babel.config.js
+++ b/babel.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-	"presets": [ "@babel/env" ]
+	"presets": [ "@babel/preset-env" ]
 };