--- a/benchmark/run-benchmark.js
+++ b/benchmark/run-benchmark.js
@@ -109,6 +109,18 @@ runBenchmark(
 
 );
 
+
+geometry.computeBoundsTree();
+intersectGeometry.computeBoundsTree();
+runBenchmark(
+
+	'Distance cast',
+	() => mesh.geometry.boundsTree.distancecast( mesh, intersectGeometry, geomMat, 1 ),
+	3000
+
+);
+
+
 console.log( '' );
 
 geometry.computeBoundsTree();