--- a/src/MeshBVH.js
+++ b/src/MeshBVH.js
@@ -265,4 +265,16 @@ export default class MeshBVH {
 
 	}
 
+	distancecast( mesh, geom, matrix, thresold ) {
+
+		for ( const root of this._roots ) {
+
+			if ( root.distancecast( mesh, geom, matrix, thresold ) ) return true;
+
+		}
+
+		return false;
+
+	}
+
 }