#### Example
http://raw.githack.com/gkjohnson/three-mesh-bvh/distance-functions/example/bundle/distancecast.html

#### TODO
- ~Add function to obb that early out if within the threshold distance~
- ~Add button to show closest distance (this will be slower to find)~
- ~Add a point distance function to mesh / threshold and use that to generate the volume~
- ~Fix LGTM issues~
- ~Change distanceToPoint to return the actual distance to the point so the isosurface parameter can be used with marching cubes.~ (Eh not quite worth it. It's fast to regenerate the volume)
- ~Change `distancecast` to `closestPointToGeometry` (others should update, as well)~

#### Make Issues
- Adding a sphere to the triangle class affords better performance in some places but worse in others. Find a way to improve the caching and only update what is needed.

#### Benchmarks

Note that a bug was fixed with the Geometrycast which impacted the performance.

##### Before
```
Compute Bounds Tree      : 218.642857 ms
Default Raycast          : 29.742574 ms
BVH Raycast              : 0.129545 ms
First Hit Raycast        : 0.024801 ms
Spherecast               : 0.005534 ms
Boxcast                  : 0.000844 ms
Geometrycast without BVH : 22.440299 ms
Geometrycast with BVH    : 7.668367 ms

BVH Memory Usage         : 10067.068 kb
```

##### After
```
Compute Bounds Tree      : 218.642857 ms
Default Raycast          : 28.961538 ms
BVH Raycast              : 0.130045 ms
First Hit Raycast        : 0.024136 ms

Spherecast               : 0.005734 ms
Boxcast                  : 0.000906 ms
Geometrycast without BVH : 859.5 ms
Geometrycast with BVH    : 1.238646 ms

Distance cast            : 0.006805 ms
Distance To Point        : 0.001204 ms

BVH Memory Usage         : 10067.068 kb
```