@gkjohnson 2019-03-09
This pull request **introduces 3 alerts** when merging 1df93b48d0014ed33f52a0c911e2aec83b4d889d into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-ee3979a56c61a7ceaf697aba321dcf40539f5095)

**new alerts:**

* 1 for Unreachable statement
* 1 for Syntax error
* 1 for Unused variable, import, function or class

---

*Comment posted by [LGTM.com](https://lgtm.com)*