@gkjohnson 2019-03-09
This pull request **introduces 3 alerts** when merging bf888d06945db5855b4bdf86234dac635ae1f417 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-36ba2af8897a82e5aa0c2cfe38d72e79c7359f8f)

**new alerts:**

* 1 for Unreachable statement
* 1 for Syntax error
* 1 for Unused variable, import, function or class

---

*Comment posted by [LGTM.com](https://lgtm.com)*