@gkjohnson 2019-03-09
This pull request **introduces 4 alerts** when merging 78a57783111896064fb155d571d2f2ff11959e83 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-512f25faac18e4f3a2540bac6106629a4966430f)

**new alerts:**

* 1 for Useless assignment to local variable
* 1 for Unreachable statement
* 1 for Assignment to constant
* 1 for Syntax error

---

*Comment posted by [LGTM.com](https://lgtm.com)*