@gkjohnson 2019-03-09
This pull request **introduces 20 alerts** when merging 95f5257f66709c1f00ef4ff7c902e8d7f072ed9c into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-f6a9ad6cdc9bdbb28645fd320c63ff8e4f7fe82e)

**new alerts:**

* 14 for Unneeded defensive code
* 3 for Unused variable, import, function or class
* 1 for Useless assignment to local variable
* 1 for Unreachable statement
* 1 for Assignment to constant

---

*Comment posted by [LGTM.com](https://lgtm.com)*