@gkjohnson 2019-03-09
This pull request **introduces 3 alerts** when merging e209e08f238eae9c30a1d6db7c8e26c6b57b2209 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-a3076babd0802d1c654c9d1d9c1f99d53bb4e20b)

**new alerts:**

* 1 for Useless assignment to local variable
* 1 for Assignment to constant
* 1 for Unused variable, import, function or class

---

*Comment posted by [LGTM.com](https://lgtm.com)*