@gkjohnson 2019-03-09
This pull request **introduces 1 alert** when merging 0612e4dea06237192c8375a945c0717eed4ed164 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-370db7c36ae1a203be13350c114fcbaa88a8eba9)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*