@gkjohnson 2019-03-09
This pull request **introduces 1 alert** when merging ae0e539bc9eb421e22f60ab572cfe692506d2fa9 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-b521a751fb51fd3bd6608cd6f530129d76f11c32)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*