@gkjohnson 2019-03-09
This pull request **introduces 1 alert** when merging fb5b1775772cd9796e463cbe31d640a05b3ec054 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-09c079a1e283d231c9c3884101884d1ffe381c73)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*