@gkjohnson 2019-03-09
This pull request **introduces 1 alert** when merging 47436d9f24f9ccd966923435f6976d82936af4f2 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-93c0b7a92047ac6fe686f5a0a4f17f8e16b8fa8e)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*