@gkjohnson 2019-03-09
This pull request **introduces 1 alert** when merging e2619d602b297e2c320854d66bf32ff5f62481ff into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-600c5b3f9ecaa5ee9e3763534f9eef7c9eb5e75f)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*