@gkjohnson 2019-03-09
This pull request **introduces 1 alert** when merging e4356c74094a5163fad09b5713a89f08d70e7747 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-0b89ec723a7a29adf6c619beb138e484baf6e219)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*