@gkjohnson 2019-03-09
This pull request **introduces 1 alert** when merging 7ee7608d38b8b1fcbabe659c405a514fde9a67c7 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-87b1b1a87e7289871f2cd4bf3c10bdc0cf6fe23f)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*