@gkjohnson 2019-03-10
This pull request **introduces 1 alert** when merging a19a3e37b10b9327c55ba08ffafe5e953c9f3669 into 32129a4229cfd47efc28116f662f8f72d247693d - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-a605ceda431aa1d2aeb5745ddde1a65ee3123ef9)

**new alerts:**

* 1 for Useless assignment to local variable

---

*Comment posted by [LGTM.com](https://lgtm.com)*